custom-elements/registering-custom-elements/unresolved-elements-interface-test-001.html

<!DOCTYPE html>
<html>
<head>
<title>Unresolved element interface must be HTMLElement, if the namespace is HTML Namespace</title>
<meta name="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<meta name="assert" content="When an unresolved element is created, it's element interface must be HTMLElement, if the namespace is HTML Namespace">
<link rel="help" href="http://www.w3.org/TR/custom-elements/#registering-custom-elements">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src='../testcommon.js'></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script type="text/javascript">

test(function() {
    var doc = newHTMLDocument();
    
    doc.body.innerHTML = '<x-foo id="xfoo"></x-foo>';

    var xfoo = doc.querySelector('#xfoo');

    assert_not_equals(xfoo, null, 'Unregistered custom element should not be null');
    
    assert_class_string(xfoo, 'HTMLElement',
            'Unregistered custom element must be a HTML element');
    

}, 'Unresolved element interface with valid name for HTML namespace');


test(function() {
    var doc = newHTMLDocument();

    var xfoo = doc.createElement('x-foo-2');
    
    assert_class_string(xfoo, 'HTMLElement',
            'Element must be a HTML element');
    
}, 'Unresolved element interface with valid name for HTML namespace. '
    + 'Element created by createElement');


test(function() {
    var doc = newHTMLDocument();

    var xfoo = doc.createElementNS(HTML_NAMESPACE, 'x-foo-3');
    
    assert_class_string(xfoo, 'HTMLElement',
            'Element must be a HTML element');
    
}, 'Unresolved element interface with valid name for HTML namespace. '
    + 'Element created by createElementNS');


test(function() {
    var doc = newHTMLDocument();
    
    doc.body.innerHTML = '<xfoo id="xfoo"></xfoo>';

    var xfoo = doc.querySelector('#xfoo');

    assert_not_equals(xfoo, null, 'Element should not be null');
    
    assert_class_string(xfoo, 'HTMLUnknownElement',
            'Element must be a HTML Unknown element');
    

}, 'Unresolved element interface with invalid name for HTML namespace');


test(function() {
    var doc = newHTMLDocument();

    var xfoo = doc.createElement('xfoo2');
    
    assert_class_string(xfoo, 'HTMLUnknownElement',
            'Element must be a HTML Unknown element');
    
}, 'Unresolved element interface with invalid name for HTML namespace. '
    + 'Element created by createElement');


test(function() {
    var doc = newHTMLDocument();

    var xfoo = doc.createElementNS(HTML_NAMESPACE, 'xfoo3');
    
    assert_class_string(xfoo, 'HTMLUnknownElement',
            'Element must be a HTML Unknown element');
    
}, 'Unresolved element interface with invalid name for HTML namespace. '
    + 'Element created by createElementNS');

</script>
</body>
</html>
